Remove trailing 0 from count

+++ db/migrate/20211114120311_add_count_ratings_to_movies.rb
@@ -1,6 +1,6 @@
 class AddCountRatingsToMovies < ActiveRecord::Migration[6.0]
   def change
-    add_column :movies, :count, :decimal
+    add_column :movies, :count, :integer
     add_index :movies, :count
     add_column :movies, :rating, :decimal
     add_index :movies, :rating
